refactor: supertest test files require listening to own server
