+++ server/package.json
@@ -21,5 +21,9 @@
     "multer": "^1.4.4",
     "nodemon": "^2.0.15",
     "pg": "^8.7.1"
+  },
+  "devDependencies": {
+    "jest": "^27.4.7",
+    "supertest": "^6.2.2"
   }
 }
